% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Add setlx engine to knitr.}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{Not used.}

\item{pkgname}{Not used.}
}
\value{
Nothing (invisible)
}
\description{
See https://randoom.org/Software/SetlX/
}
\details{
Invoke the setlx engine like so:

\if{html}{\out{<div class="sourceCode">}}\preformatted{```\{setlx\}
...
```
}\if{html}{\out{</div>}}
}
\author{
fnaufel
}
