% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{_PACKAGE}
\name{rsetlx}
\alias{rsetlx}
\title{Add setlx engine to knitr.}
\description{
See https://randoom.org/Software/SetlX/
}
\details{
Invoke the setlx engine like so:

\if{html}{\out{<div class="sourceCode">}}\preformatted{```\{setlx\}
...
```
}\if{html}{\out{</div>}}
}
\author{
fnaufel
}
